src/test/java/org/example/solutions/stack/EvaluateReversePolishNotationTest.java
package org.example.solutions.stack;

import java.util.stream.Stream;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

class EvaluateReversePolishNotationTest {

    @ParameterizedTest
    @MethodSource("testCases")
    void testSolution(String[] nums, int expectedResult) {
        var solution = new EvaluateReversePolishNotation();
        var result = solution.evalRPN(nums);
        Assertions.assertEquals(expectedResult, result);
    }

    public static Stream<Arguments> testCases() {

        return Stream.of(
                Arguments.of(new String[]{"2", "1", "+", "3", "*"}, 9),
                Arguments.of(new String[]{"4", "13", "5", "/", "+"}, 6),
                Arguments.of(new String[]{"10", "6", "9", "3", "+", "-11", "*", "/", "*", "17", "+", "5", "+"}, 22)
                );
    }

}